Update Ride Creation Logic and User Information
Users may now update cars with a picture.

Rides now have a relationship with a model named RidePickups which serves as a reference to pickups a driver may have for a ride. This will be displayed as the number of "stops" a ride has.

Navigation links and headers have been updated.

Single Ride page now displays the correct ride information

Fixed a bug where rides were not saved with the pickup price determined by the driver.

Fixed a bug where user avatars were not saving to the correct location