Fix Single Ride Errors, Phone Carrier Update, Join AS Driver
Fixed a Bug where users could not join a ride as a driver due to uncomplete input form

Fixed the logic that determined if the logged in user can join a ride as a passenger or driver

Fixed the logic that allowed users to join a ride as a passenger

Updated Passenger and Driver Ride Request functions

Added a function to determine the carrier of a users phone upon registration

Added Completed Request/ Submit Form validation and Notifications

Fixed a bug where Ride Slugs were saved with the date appended to the end of the string which caused unexpected erros relating to the url to the single Ride.

Removed Ride Locations

Updated Single Ride Styles
Added Minutes To Hours Time Conversion

Three items are still due for completion and may not be included in the beta test

Transactions with Account Balances
Confirmation of Rides
Ride Reviews.